fix(confg): gen_config

+++ src/config.rs
@@ -1,18 +1,19 @@
 use anyhow::{Context, Error, Result};
 use directories::ProjectDirs;
 use serde::{Deserialize, Serialize};
-use std::{fs, path::PathBuf};
+use std::{env, fs, path::PathBuf};
 
 const CONFIG: &str = "config.toml";
 const APP_NAME: &str = "psu";
 
-#[derive(Deserialize, Serialize, Debug)]
+#[derive(Deserialize, Serialize, Debug, Default)]
 pub struct Config {
-	fields: Fields,
+	pub path: PathBuf,
+	pub fields: Fields,
 }
 
 #[derive(Deserialize, Default, Serialize, Debug)]
-struct Fields {
+pub struct Fields {
 	input: String,
 	login: String,
 	service: String,
@@ -21,21 +22,21 @@ struct Fields {
 impl Config {
 	pub fn new() -> Self {
 		Self {
-			fields: Fields::default(),
+			path: env::home_dir().unwrap_or_default(),
+			..Default::default()
 		}
 	}
 
-	pub fn gen_config(&self) -> Result<()> {
+	pub fn gen_config() -> Result<()> {
 		let dir = get_app_data_dir().context("Failed to obtain config directory.")?;
-		let conf = self.parse_config().context("Failed to parse configuration.")?;
 
-		fs::create_dir_all(&dir).context("Failed create config dirs")?;
-		fs::write(dir.join(CONFIG), conf).context("Failed write config.")?;
+		if !dir.exists() {
+			fs::create_dir_all(&dir).context("Failed create config dirs")?;
+			let conf = toml::to_string_pretty(&Config::new()).context("Failed to parse configuration.")?;
+			fs::write(dir.join(CONFIG), conf).context("Failed write config.")?;
+		}
+
 		Ok(())
-	}
-
-	fn parse_config(&self) -> Result<String, Error> {
-		Ok(toml::to_string(self)?)
 	}
 }
